Edit CI job ("Build Documents")
Edited the CI job "Build Documents" to build on push events for
all branches except the 'gh-pages' branch. The output of each
build is then stored in the 'gh-pages' branch, in the main folder
(if the current branch is 'master') or in a folder (contained into
the 'branches/' folder)  with the same name as the current branch.